+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # master
 
+* Display an error message if bundler failed to run or is not present in the system
+
+    * Juan Manuel Ramallo*
+
 # 0.1.3
 
 * Fix gem creation task on Linux